backend/requirements.txt
Flask
pymongo[srv]
python-dotenv
flask-cors
bcrypt
matplotlib
cloudinary
apscheduler
emails
pandas
requests